feat(agent): add startAgentJob method to AgentService and update job service to utilize it
